+++ CliError.test.mjs
@@ -1,12 +1,12 @@
 // @ts-check
 
 import { strictEqual, throws } from "node:assert";
-import { describe, it } from "node:test";
+import { suite, test } from "node:test";
 
 import CliError from "./CliError.mjs";
 
-describe("Class `CliError`.", { concurrency: true }, () => {
-  it("Argument 1 `message` not a string.", () => {
+suite("Class `CliError`.", { concurrency: true }, () => {
+  test("Argument 1 `message` not a string.", () => {
     throws(() => {
       new CliError(
         // @ts-expect-error Testing invalid.
@@ -15,7 +15,7 @@ describe("Class `CliError`.", { concurrency: true }, () => {
     }, new TypeError("Argument 1 `message` must be a string."));
   });
 
-  it("Arguments valid.", () => {
+  test("Arguments valid.", () => {
     const message = "Message.";
     const error = new CliError(message);