+++ directoryPathToFileURL.test.mjs
@@ -1,12 +1,12 @@
 // @ts-check
 
 import { deepStrictEqual, throws } from "node:assert";
-import { describe, it } from "node:test";
+import { suite, test } from "node:test";
 
 import directoryPathToFileURL from "./directoryPathToFileURL.mjs";
 
-describe("Function `directoryPathToFileURL`.", { concurrency: true }, () => {
-  it("Argument 1 `directoryPath` not a string.", () => {
+suite("Function `directoryPathToFileURL`.", { concurrency: true }, () => {
+  test("Argument 1 `directoryPath` not a string.", () => {
     throws(() => {
       directoryPathToFileURL(
         // @ts-expect-error Testing invalid.
@@ -15,7 +15,7 @@ describe("Function `directoryPathToFileURL`.", { concurrency: true }, () => {
     }, new TypeError("Argument 1 `directoryPath` must be a string."));
   });
 
-  it("Directory path ends with `/`.", () => {
+  test("Directory path ends with `/`.", () => {
     const directoryPath = "/a/b/c/";
 
     deepStrictEqual(
@@ -24,7 +24,7 @@ describe("Function `directoryPathToFileURL`.", { concurrency: true }, () => {
     );
   });
 
-  it("Directory path doesn’t end with `/`.", () => {
+  test("Directory path doesn’t end with `/`.", () => {
     const directoryPath = "/a/b/c";
 
     deepStrictEqual(